Fix showing tooltips with non-zero line spacing
x-max-tooltip-size uses rows, but it seems that they are internally
converted to pixels using character height, and if your text doesn't fit
it gets truncated.

This works as expected with no line spacing, but with non-zero spacing
the result of division will be a float that would previously get
truncated leading subsequently to smaller maximum height, cutting off the
last line of the tooltip.

So, round the row count up. Fixes #19.